engineapi: Introduce cmd for frontend start and shutdown
